Refactor
Api divided to different modules: user, offers and reviws
